+++ plugins/groovy/src/org/jetbrains/plugins/groovy/intentions/style/inference/MethodParametersInferenceProcessor.kt
@@ -1,26 +1,13 @@
 // Copyright 2000-2019 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 package org.jetbrains.plugins.groovy.intentions.style.inference
 
-import com.intellij.psi.*
-import com.intellij.psi.search.searches.ReferencesSearch
-import com.intellij.util.containers.toArray
-import org.jetbrains.plugins.groovy.lang.psi.GroovyFile
-import org.jetbrains.plugins.groovy.lang.psi.GroovyPsiElementFactory
-import org.jetbrains.plugins.groovy.lang.psi.GroovyRecursiveElementVisitor
-import org.jetbrains.plugins.groovy.lang.psi.api.GroovyMethodResult
-import org.jetbrains.plugins.groovy.lang.psi.api.statements.expressions.GrExpression
-import org.jetbrains.plugins.groovy.lang.psi.api.statements.expressions.GrOperatorExpression
-import org.jetbrains.plugins.groovy.lang.psi.api.statements.expressions.path.GrCallExpression
-import org.jetbrains.plugins.groovy.lang.psi.api.statements.params.GrParameter
+import com.intellij.psi.PsiIntersectionType
+import com.intellij.psi.PsiSubstitutor
+import com.intellij.psi.PsiType
+import com.intellij.psi.PsiWildcardType
 import org.jetbrains.plugins.groovy.lang.psi.api.statements.typedef.members.GrMethod
-import org.jetbrains.plugins.groovy.lang.psi.util.GroovyCommonClassNames
-import org.jetbrains.plugins.groovy.lang.resolve.processors.inference.ExpressionConstraint
 import org.jetbrains.plugins.groovy.lang.resolve.processors.inference.GroovyInferenceSession
-import org.jetbrains.plugins.groovy.lang.resolve.processors.inference.OperatorExpressionConstraint
 import org.jetbrains.plugins.groovy.lang.resolve.processors.inference.type
-import java.util.*
-import kotlin.collections.HashMap
-import kotlin.collections.set
 
 
 /**
@@ -30,13 +17,7 @@ import kotlin.collections.set
 /**
  * Allows to infer method parameters types regarding method calls and inner dependencies between types.
  */
-class MethodParametersInferenceProcessor(val method: GrMethod) {
-  private val elementFactory: GroovyPsiElementFactory = GroovyPsiElementFactory.getInstance(method.project)
-  private val defaultTypeParameterList = (method.typeParameterList?.copy()
-                                          ?: elementFactory.createTypeParameterList()) as PsiTypeParameterList
-  private val nameGenerator = NameGenerator()
-  private val targetParameters = method.parameters.filter { it.typeElement == null }
-  private val closureTypeParameters: MutableMap<GrParameter, ParametrizedClosure> = HashMap()
+class MethodParametersInferenceProcessor(private val method: GrMethod, private val driver: InferenceDriver = InferenceDriverImpl(method)) {
 
   /**
    * Performs full substitution for non-typed parameters of [method]
@@ -46,64 +27,38 @@ class MethodParametersInferenceProcessor(val method: GrMethod) {
    * 3. Inferring dependencies between new type parameters and instantiating them.
    */
   fun runInferenceProcess() {
-    val parameterIndex = setUpNewTypeParameters()
-    val constantTypeParameters = setUpParametersSignature(parameterIndex)
-    val registry = setUpRegistry(constantTypeParameters)
+    driver.setUpNewTypeParameters()
+    setUpParametersSignature()
+    val registry = setUpRegistry()
     inferTypeParameters(registry)
-    if (method.typeParameters.isEmpty()) {
-      method.typeParameterList?.delete()
-    }
   }
 
-
-  /**
-   * Does first phase of inference
-   */
-  private fun setUpNewTypeParameters(): Map<GrParameter, PsiTypeParameter> {
-    if (!method.hasTypeParameters()) {
-      method.addAfter(elementFactory.createTypeParameterList(), method.firstChild)
-    }
-    val typeParameterList = method.typeParameterList!!
-    val parameterIndex = LinkedHashMap<GrParameter, PsiTypeParameter>()
-
-    val generator = NameGenerator()
-    for (parameter in method.parameters.filter { it.typeElement == null }) {
-      val newTypeParameter = createNewTypeParameter(generator)
-      typeParameterList.add(newTypeParameter)
-      parameterIndex[parameter] = newTypeParameter
-      parameter.setType(newTypeParameter.type())
-    }
-    createParametrizedClosures(method, closureTypeParameters, this, generator)
-    return parameterIndex
-  }
-
-  fun createNewTypeParameter(generator: NameGenerator? = null): PsiTypeParameter {
-    return elementFactory.createProperTypeParameter((generator ?: nameGenerator).name, PsiClassType.EMPTY_ARRAY)
-  }
-
-  /**
-   * Does second phase of inference
-   */
-  private fun setUpParametersSignature(parameterIndex: Map<GrParameter, PsiTypeParameter>): Array<PsiTypeParameter> {
+  private fun setUpParametersSignature() {
     val inferenceSession = GroovyInferenceSession(method.typeParameters, PsiSubstitutor.EMPTY, method,
                                                   propagateVariablesToNestedSessions = true)
-    collectOuterMethodCalls(inferenceSession)
-    closureTypeParameters.values.forEach { setUpClosuresSignature(inferenceSession, it) }
+    driver.collectOuterCalls(inferenceSession)
     val signatureSubstitutor = inferenceSession.inferSubst()
-    method.typeParameterList?.replace(defaultTypeParameterList.copy())
-    val defaultTypeParameters = method.typeParameters
-    for ((parameter, typeParameter) in parameterIndex) {
-      parameter.setType(
-        createDeeplyParametrizedType(signatureSubstitutor.substitute(typeParameter)!!, method.typeParameterList!!,
-                                     parameter, signatureSubstitutor))
-    }
-    return defaultTypeParameters
+    driver.parametrizeMethod(signatureSubstitutor)
   }
 
 
-  /**
-   * Does third phase of inference
-   */
+  private fun setUpRegistry(): InferenceUnitRegistry {
+    val inferenceSession = GroovyInferenceSession(method.typeParameters, PsiSubstitutor.EMPTY, method,
+                                                  propagateVariablesToNestedSessions = true)
+    driver.collectInnerMethodCalls(inferenceSession)
+    driver.constantTypes.forEach { getInferenceVariable(inferenceSession, it).instantiation = it }
+    inferenceSession.run { repeatInferencePhases(); infer() }
+    val inferenceVariables = method.typeParameters.map { getInferenceVariable(inferenceSession, it.type()) }
+    return InferenceUnitRegistry().apply {
+      setUpUnits(inferenceVariables, inferenceSession)
+      driver.constantTypes.forEach { searchUnit(it)?.constant = true }
+      driver.flexibleTypes.forEach { searchUnit(it)?.flexible = true }
+      driver.forbiddingTypes.mapNotNull { searchUnit(it) }
+        .filter { it.typeInstantiation == PsiType.getJavaLangObject(method.manager, method.resolveScope) }
+        .forEach { it.forbidInstantiation = true }
+    }
+  }
+
   private fun inferTypeParameters(registry: InferenceUnitRegistry) {
     val graph = InferenceUnitGraph(registry)
     val representativeSubstitutor = collectRepresentativeSubstitutor(graph, registry)
@@ -112,24 +67,7 @@ class MethodParametersInferenceProcessor(val method: GrMethod) {
       val preferableType = getPreferableType(graph, unit, representativeSubstitutor, resultSubstitutor)
       graph.getEqualUnits(unit).forEach { resultSubstitutor = resultSubstitutor.put(it.initialTypeParameter, preferableType) }
     }
-    if (targetParameters.any { isClosureType(it.type) }) {
-      ParametrizedClosure.ensureImports(elementFactory, method.containingFile as GroovyFile)
-    }
-    targetParameters.forEach { param ->
-      param.setType(resultSubstitutor.substitute(param.type))
-      when {
-        isClosureType(param.type) -> {
-          closureTypeParameters[param]!!.run {
-            substituteTypes(resultSubstitutor)
-            renderTypes(elementFactory)
-          }
-        }
-        param.type is PsiArrayType -> param.setType(
-          registry.searchUnit((param.type as PsiArrayType).componentType)!!.type.createArrayType())
-      }
-    }
-    val residualTypeParameterList = buildResidualTypeParameterList(defaultTypeParameterList)
-    method.typeParameterList?.replace(residualTypeParameterList)
+    driver.acceptFinalSubstitutor(resultSubstitutor)
   }
 
   private fun getPreferableType(graph: InferenceUnitGraph,
@@ -154,223 +92,10 @@ class MethodParametersInferenceProcessor(val method: GrMethod) {
       else -> {
         val parent = unit.unitInstantiation
         val advice = if (parent?.typeInstantiation == PsiType.NULL) parent?.type else parent?.typeInstantiation
-        val newTypeParam = createBoundedTypeParameterElement(unit, representativeSubstitutor, advice)
-        defaultTypeParameterList.add(newTypeParam)
+        val newTypeParam = driver.createBoundedTypeParameterElement(unit.initialTypeParameter.name!!, representativeSubstitutor,
+                                                                    advice ?: unit.typeInstantiation)
         return newTypeParam.type()
       }
     }
   }
-
-  private fun setUpRegistry(constantTypeParameters: Array<PsiTypeParameter>): InferenceUnitRegistry {
-    val inferenceSession = GroovyInferenceSession(method.typeParameters, PsiSubstitutor.EMPTY, method,
-                                                  propagateVariablesToNestedSessions = true)
-    collectInnerMethodCalls(inferenceSession)
-    constantTypeParameters.map { it.type() }.forEach { getInferenceVariable(inferenceSession, it).instantiation = it }
-    inferenceSession.run { repeatInferencePhases(); infer() }
-    val inferenceVariables = method.typeParameters.map { getInferenceVariable(inferenceSession, it.type()) }.toList()
-    return InferenceUnitRegistry().apply {
-      setUpUnits(inferenceVariables, inferenceSession)
-      constantTypeParameters.map { it.type() }.forEach { searchUnit(it)?.constant = true }
-      method.parameters.run {
-        mapNotNull { searchUnit(it.type) }.forEach { it.flexible = true }
-        mapNotNull { searchUnit((it.type as? PsiArrayType)?.componentType) }
-          .filter { it.typeInstantiation == PsiType.getJavaLangObject(method.manager, method.resolveScope) }
-          .forEach { it.forbidInstantiation = true }
-      }
-      // todo: incorrect flexibility
-      closureTypeParameters.values
-        .flatMap { it.typeParameters }
-        .forEach { searchUnit(it.type())!!.flexible = true }
-    }
-  }
-
-
-  /**
-   * Creates type parameter with upper bound of [target].
-   * If [target] is parametrized, all it's parameter types will also be parametrized.
-   */
-  private fun createDeeplyParametrizedType(target: PsiType,
-                                           typeParameterList: PsiTypeParameterList,
-                                           parameter: GrParameter?,
-                                           signatureSubstitutor: PsiSubstitutor
-  ): PsiType {
-    val visitor = object : PsiTypeMapper() {
-
-      private fun registerTypeParameter(vararg supertypes: PsiClassType): PsiType {
-        val typeParameter = elementFactory.createProperTypeParameter(nameGenerator.name, supertypes.filter {
-          !it.equalsToText(GroovyCommonClassNames.GROOVY_OBJECT)
-        }.toTypedArray())
-        typeParameterList.add(typeParameter)
-        return typeParameter.type()
-      }
-
-      override fun visitClassType(classType: PsiClassType?): PsiType? {
-        classType ?: return classType
-        val generifiedClassType = if (classType.isRaw) {
-          val resolveResult = classType.resolve()!!
-          val wildcards = Array(resolveResult.typeParameters.size) { PsiWildcardType.createUnbounded(method.manager) }
-          elementFactory.createType(resolveResult, *wildcards)
-        }
-        else classType
-        val parameters = generifiedClassType.parameters
-        val replacedParameters = parameters.map { it.accept(this) }.toArray(emptyArray())
-        val resolveResult = generifiedClassType.resolveGenerics()
-        if (classType == PsiType.getJavaLangObject(method.manager, method.resolveScope)) {
-          return registerTypeParameter()
-        }
-        else {
-          return elementFactory.createType(resolveResult.element ?: return null, *replacedParameters)
-        }
-      }
-
-      override fun visitWildcardType(wildcardType: PsiWildcardType?): PsiType? {
-        wildcardType ?: return wildcardType
-        val upperBounds = if (wildcardType.isExtends) arrayOf(wildcardType.extendsBound.accept(this) as PsiClassType) else emptyArray()
-        return registerTypeParameter(*upperBounds)
-      }
-
-      override fun visitCapturedWildcardType(type: PsiCapturedWildcardType?): PsiType? {
-        type ?: return type
-        val upperBound = type.upperBound.accept(this) as PsiClassType
-        return registerTypeParameter(upperBound)
-      }
-
-      override fun visitIntersectionType(intersectionType: PsiIntersectionType?): PsiType? {
-        intersectionType ?: return intersectionType
-        val parametrizedConjuncts = intersectionType.conjuncts.map { it.accept(this) as PsiClassType }.toTypedArray()
-        return registerTypeParameter(*parametrizedConjuncts)
-      }
-
-    }
-    when {
-      isClosureType(target) -> {
-        val closureParameter = closureTypeParameters[parameter]!!
-        val initialTypeParameters = typeParameterList.typeParameters
-        closureParameter.typeParameters.run {
-          forEach {
-            val topLevelType = createDeeplyParametrizedType(signatureSubstitutor.substitute(it)!!, typeParameterList, null,
-                                                            signatureSubstitutor)
-            closureParameter.types.add(topLevelType)
-          }
-          val createdTypeParameters = typeParameterList.typeParameters.subtract(initialTypeParameters.asIterable())
-          clear()
-          addAll(createdTypeParameters)
-        }
-        return target.accept(visitor)
-      }
-      target is PsiIntersectionType || target is PsiClassType && (target.hasParameters() || target.isRaw) -> return target.accept(visitor)
-      target is PsiArrayType -> {
-        val newTypeParameter = elementFactory.createProperTypeParameter(nameGenerator.name, emptyArray())
-        typeParameterList.add(newTypeParameter)
-        return newTypeParameter.type().createArrayType()
-      }
-      else -> {
-        val typeParam = elementFactory.createProperTypeParameter(nameGenerator.name, arrayOf(target as PsiClassType))
-        typeParameterList.add(typeParam)
-        return typeParam.type()
-      }
-    }
-  }
-
-
-  private fun buildResidualTypeParameterList(typeParameterList: PsiTypeParameterList): PsiTypeParameterList {
-    method.typeParameterList!!.replace(typeParameterList)
-    val necessaryTypeParameters = LinkedHashSet<PsiTypeParameter>()
-    val visitor = object : PsiTypeVisitor<PsiType>() {
-      override fun visitClassType(classType: PsiClassType?): PsiType? {
-        classType ?: return classType
-        val resolveElement = classType.resolveGenerics().element
-        if (resolveElement is PsiTypeParameter) {
-          if (resolveElement.text !in necessaryTypeParameters.map { it.text }) {
-            necessaryTypeParameters.add(resolveElement)
-          }
-          resolveElement.extendsList.referencedTypes.forEach { it.accept(this) }
-        }
-        classType.parameters.forEach { it.accept(this) }
-        return super.visitClassType(classType)
-      }
-
-      override fun visitWildcardType(wildcardType: PsiWildcardType?): PsiType? {
-        wildcardType?.extendsBound?.accept(this)
-        return super.visitWildcardType(wildcardType)
-      }
-
-      override fun visitCapturedWildcardType(capturedWildcardType: PsiCapturedWildcardType?): PsiType? {
-        capturedWildcardType?.wildcard?.accept(this)
-        return super.visitCapturedWildcardType(capturedWildcardType)
-      }
-
-      override fun visitArrayType(arrayType: PsiArrayType?): PsiType? {
-        arrayType?.componentType?.accept(this)
-        return super.visitArrayType(arrayType)
-      }
-    }
-    method.parameters.forEach { it.type.accept(visitor) }
-    closureTypeParameters.values.flatMap { it.types }.forEach { it.accept(visitor) }
-    val resultingTypeParameterList = elementFactory.createTypeParameterList()
-    necessaryTypeParameters.forEach { resultingTypeParameterList.add(it) }
-    return resultingTypeParameterList
-  }
-
-  /**
-   * Creates ready for insertion type parameter with correct extends list.
-   */
-  private fun createBoundedTypeParameterElement(unit: InferenceUnit,
-                                                relationSubstitutor: PsiSubstitutor = PsiSubstitutor.EMPTY,
-                                                supertypeAdvice: PsiType? = null): PsiTypeParameter {
-    val superType = supertypeAdvice ?: unit.typeInstantiation
-    val mappedSupertypes = when (superType) {
-      is PsiClassType -> arrayOf(relationSubstitutor.substitute(superType) as PsiClassType)
-      is PsiIntersectionType -> superType.conjuncts.map { relationSubstitutor.substitute(it) as PsiClassType }.toTypedArray()
-      else -> emptyArray()
-    }
-    return elementFactory.createProperTypeParameter(unit.initialTypeParameter.name!!, mappedSupertypes)
-  }
-
-
-  /**
-   * Scans [method] body for calls and registers dependencies between type parameters.
-   */
-  private fun collectInnerMethodCalls(resolveSession: GroovyInferenceSession) {
-    method.accept(object : GroovyRecursiveElementVisitor() {
-      override fun visitCallExpression(callExpression: GrCallExpression) {
-        val receiver = (callExpression.advancedResolve() as? GroovyMethodResult)?.candidate?.receiver as? PsiClassType
-        if (isClosureType(receiver)) {
-          val parameter = callExpression.firstChild.run {
-            closureTypeParameters[reference?.resolve() as? GrParameter ?: firstChild.reference?.resolve()]
-          }
-          parameter?.run {
-            callExpression.expressionArguments.zip(parameter.types).forEach { (expression, parameterType) ->
-              resolveSession.addConstraint(ExpressionConstraint(resolveSession.substituteWithInferenceVariables(parameterType), expression))
-            }
-          }
-        }
-        else {
-          resolveSession.addConstraint(ExpressionConstraint(null, callExpression))
-        }
-        super.visitCallExpression(callExpression)
-      }
-
-      override fun visitExpression(expression: GrExpression) {
-        if (expression is GrOperatorExpression) {
-          resolveSession.addConstraint(OperatorExpressionConstraint(expression))
-        }
-        super.visitExpression(expression)
-      }
-    })
-  }
-
-  /**
-   * Scans environment for calls of [method] to infer basic parameter signature.
-   */
-  private fun collectOuterMethodCalls(resolveSession: GroovyInferenceSession) {
-    val references = ReferencesSearch.search(method).findAll()
-    for (parameter in method.parameters) {
-      resolveSession.addConstraint(ExpressionConstraint(parameter.type, parameter.initializerGroovy ?: continue))
-    }
-    for (call in references.mapNotNull { it.element.parent as? GrExpression }) {
-      resolveSession.addConstraint(ExpressionConstraint(null, call))
-    }
-  }
-
 }
